requirements.txt
httpx>=0.27.0
pyyaml>=6.0
rich>=13.0
